meal_planner/core/ai/recommenders/meal_recommender.py
import pandas as pd
import numpy as np
from sklearn.preprocessing import MinMaxScaler
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import LSTM, Dense, Dropout

class MealRecommender:
    def __init__(self, data_path):
        self.df = pd.read_csv(data_path)
        self.scaler = MinMaxScaler()
        self._preprocess_data()
        
    def _preprocess_data(self):
        # Add dummy pricing data
        self.df['Price'] = np.random.uniform(1, 20, len(self.df))
        
        # Nutritional features
        self.features = self.df[[
            'Calories (kcal)', 'Protein (g)', 
            'Carbohydrates (g)', 'Fat (g)', 'Price'
        ]]
        self.scaled_features = self.scaler.fit_transform(self.features)
        
    def build_lstm_model(self):
        model = Sequential([
            LSTM(128, return_sequences=True, input_shape=(7, 5)),
            Dropout(0.2),
            LSTM(64),
            Dense(64, activation='relu'),
            Dense(len(self.df), activation='softmax')
        ])
        model.compile(loss='categorical_crossentropy', optimizer='adam')
        return model
    
    def generate_plan(self, constraints):
        # Placeholder recommendation logic
        filtered = self._apply_constraints(constraints)
        return filtered.sample(3)

    def _apply_constraints(self, constraints):
        # Implement actual constraint logic
        return self.df